moved gcc methods to gcc_utils; Uploads json data directly from blog w/o temp file storage
